requirements.txt
asgiref==3.7.2
Django==4.2.5
django-environ==0.11.2
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.7.1